Fixed Float32->Float64

+++ workspace/src/barc/src/barc_record.jl
@@ -29,12 +29,12 @@ include("barc_lib/log_functions.jl")
 function main() 
 
     buffersize      = 60000
-    gps_meas        = Measurements{Float32}(1,zeros(buffersize),zeros(buffersize),zeros(buffersize,2))
-    imu_meas        = Measurements{Float32}(1,zeros(buffersize),zeros(buffersize),zeros(buffersize,9))
-    cmd_log         = Measurements{Float32}(1,zeros(buffersize),zeros(buffersize),zeros(buffersize,2))
-    cmd_pwm_log     = Measurements{Float32}(1,zeros(buffersize),zeros(buffersize),zeros(buffersize,2))
-    vel_est_log     = Measurements{Float32}(1,zeros(buffersize),zeros(buffersize),zeros(buffersize,5))
-    pos_info_log    = Measurements{Float32}(1,zeros(buffersize),zeros(buffersize),zeros(buffersize,20))
+    gps_meas        = Measurements{Float64}(1,zeros(buffersize),zeros(buffersize),zeros(buffersize,2))
+    imu_meas        = Measurements{Float64}(1,zeros(buffersize),zeros(buffersize),zeros(buffersize,9))
+    cmd_log         = Measurements{Float64}(1,zeros(buffersize),zeros(buffersize),zeros(buffersize,2))
+    cmd_pwm_log     = Measurements{Float64}(1,zeros(buffersize),zeros(buffersize),zeros(buffersize,2))
+    vel_est_log     = Measurements{Float64}(1,zeros(buffersize),zeros(buffersize),zeros(buffersize,5))
+    pos_info_log    = Measurements{Float64}(1,zeros(buffersize),zeros(buffersize),zeros(buffersize,20))
 
     # initiate node, set up publisher / subscriber topics
     init_node("barc_record")
